Deleting files and Floating PDFs
The changes I made to the permissions system means its kind of a pain to
delete the downloaded pdfs, so I added an option to long tap an attachment
and delete it.

attachments with no top-level parent (floating pdfs) are now properly (sort of)
displayed by the app instead of being missing.

Since this app was originally dropped 6 yrs ago, there's no support
for annotations, a recent addition. They were being displayed as "no name"
items. I made it so the system just ignores them to get rid of the no-name stuff,
adding annotations to the app properly is future work.